+++ src/modules/locale-storage.js
